refactor(racing-car): MVC 패턴 적용
